Edit and Delete Tasks

+++ app/Http/Controllers/Admin/TaskController.php
@@ -3,6 +3,7 @@
 namespace App\Http\Controllers\Admin;
 
 use App\Models\Task;
+use App\Models\User;
 use Illuminate\Http\Request;
 use App\Http\Controllers\Controller;
 
@@ -12,14 +13,16 @@ class TaskController extends Controller
     {
         $tasks = Task::with('user')->orderBy('due_date')->get();
 
-        return view('admin.tasks.index', compact('tasks'));
+        return view('admin.tasks.index', compact('tasks','users'));
     }
 
     public function create()
     {
         $this->authorize('create', Task::class);
 
-        return view('admin.tasks.create');
+        $users = User::where('is_admin',0)->get(['name','id']);
+
+        return view('admin.tasks.create', compact('users'));
     }
 
     public function update(Task $task)
@@ -31,7 +34,26 @@ class TaskController extends Controller
 
     public function destroy(Task $task)
     {
+        $this->authorize('delete', Task::class);
+
         $task->delete();
+
+        return redirect()->route('admin.tasks.index');
+    }
+    
+    public function store(Request $request)
+    {
+        $name = $request->description;
+        $user = $request->user;
+        $date = $request->due_date;
+
+        Task::create([
+            'description' => $name,
+            'due_date' => $date,
+            'user_id' => $user
+        ]);
+
+        return redirect()->route('admin.tasks.index');
     }
 
 }
